+++ account/src/main/java/com/devsu/prueba/account/entities/Movement.java
@@ -4,6 +4,7 @@ import com.devsu.prueba.account.entities.enums.MovementType;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -25,9 +26,9 @@ public class Movement {
     @Enumerated(EnumType.STRING)
     private MovementType movementType;
 
-    private float value;
+    private BigDecimal value;
 
-    private float balance;
+    private BigDecimal balance;
 
     @ManyToOne
     @JoinColumn(name = "accountId", nullable = false)
