+++ prueba/src/main/java/com/devsu/prueba/entities/Account.java
@@ -1,10 +1,12 @@
 package com.devsu.prueba.entities;
 
+import com.devsu.prueba.account.entities.Movement;
 import com.devsu.prueba.entities.enums.AccountType;
 import com.devsu.prueba.entities.enums.Status;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.Set;
 import java.util.UUID;
 
@@ -24,9 +26,9 @@ public class Account {
     @Enumerated(EnumType.STRING)
     private AccountType accountType;
 
-    private float initialBalance;
+    private BigDecimal initialBalance;
 
-    private float balance;
+    private BigDecimal balance;
 
     @NonNull
     @Enumerated(EnumType.STRING)
